Launch program only when pressing ON
Also, make it work without the bootloader. LCD initialization was
previously incomplete: the program would work, but only after having
been freshly uploaded to the calculator, which made me prematurely
declare success.

When running outside of the bootloader (removing the battery, then
putting it back on), nothing could be seen. This was because I was
missing a hack that KnightOS does that I didn't do. WikiTI doesn't
explain why we need to do this, but mentions that it's required.